Support commitfooter formatting
We want to enforce the existing of certain commit message footers.

We do this by introducing a formatter called

  fmt:commitfooter-FOOTER-NAME.DISAMBIGUATION-SUFFIX

Since commit message footers commonly have dashes (Signed-off-by,
Change-Id), we can't use '-' as separator between the
DISAMBIGUATION-SUFFIX and the language.

This is solved by changing the separator to '.'. This will invalidate
existing formatter configs.

Change-Id: Ibd6d878d3944746c723fc4d1f8699fe4fbc862f1